// src/app/api/generated/model/catalogo-concepto-page-response.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CatalogoConceptoResponse } from './catalogo-concepto-response';


/**
 * Pagina de conceptos del catalogo, segun los filtros pedidos
 */
export interface CatalogoConceptoPageResponse { 
    /**
     * Conceptos de esta pagina
     */
    content?: Array<CatalogoConceptoResponse>;
    /**
     * Numero de pagina devuelta, base cero
     */
    page?: number;
    /**
     * Cantidad de elementos por pagina
     */
    size?: number;
    /**
     * Cantidad total de conceptos que cumplen el filtro
     */
    totalElements?: number;
    /**
     * Cantidad total de paginas disponibles
     */
    totalPages?: number;
}